+++ engine/src/lib.rs
@@ -1,6 +1,8 @@
+// SPDX-License-Identifier: MIT
 mod bamboo;
 mod engine;
 mod english;
+pub mod event;
 mod input_method;
 pub mod spelling;
 
@@ -9,4 +11,5 @@ mod tests;
 
 pub use engine::Engine;
 pub use engine::EngineEvent;
+pub use event::{Command, EventStore, InputEvent};
 pub use input_method::InputMethod;